Redo gofmt support... again
Switch back to stdin/stdout gofmt processing. This has turned out
to be the only reliable way to preserve cursors and the viewport
position following a buffer replacement.

Unfortunately there's still a slight flicker when replacing the
buffer, but at least it's reliable.

Also: don't automatically jump to syntax errors. If the user wants
to navigate to an error, they can use the next_result/prev_result
commands.